[Improve]Separate the record and schema-change in JsonDebeziumSchemaSerializer (#279)
